requirements.txt
dataclasses-json
openai
pytest
pytest-mock
tavily-python
markdownify
mammoth 
markdownify
pandas
pdfminer.six 
python-pptx 
puremagic 
pydub 
SpeechRecognition
beautifulsoup4 
youtube-transcript-api
pymongo 
pydantic
termcolor
pexpect
replicate
fastapi
python-dotenv
python-jose
jsonschema
aiohttp
pymupdf
colorama
langchain_cohere
langchain-text-splitters
cleantext
markdown
yt_dlp
fal-client
docker
boto3
uvicorn
psutil
resource
y-py
uvicorn[standard]
qdrant-client
slowapi 
redis 
python-magic
ecdsa>=0.19.1

